requirements.txt
# Priority Protector Requirements - TESTED & WORKING
# This version has NO dependency conflicts!

# Core Flask
Flask==3.0.0
flask-cors==4.0.0

# Google Calendar API
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# HTTP requests for Slack
requests

# Environment variables
python-dotenv

# Additional utilities
Werkzeug==3.0.1

# NOTE: IBM watsonx Orchestrate integration is done via Slack channel
# No Python package needed - your agent "LabLabAI" handles it!
# Just use @Priority_Protector in your Slack workspace
